requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.12.4
Flask==2.3.3
future==0.18.3
gevent==23.9.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PySocks==1.7.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
tldextract==3.5.0
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
python-dotenv=1.0.0
